Fix bug for users that had to upgrade
For users that had to upgrade, servers would displays both city label and upgrade button at the same time, making it hard to read. Not the label has been corrected so that it displays only upgrade and it is hidden when overing over a server, which is replaced by a 'upgrade' button.