Add scroll to quick settings pop-over to prevent UI cuts
This issue is due to the fact that GTK popover does no bleed over parent window, thus if the content is wider/taller then the parent window, it'll just cut the popover view. This should be fixed in GTk4 but does not work in GTK3.

Source: https://gitlab.gnome.org/GNOME/gtk/-/issues/543